@@ -1,53 +1,46 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections;
 
-namespace RyanJuan.Hestia
+namespace RyanJuan.Hestia;
+
+public static partial class HestiaCollections
 {
-    public static partial class HestiaCollections
-    {
 #if ZH_HANT
 #else
 #endif
-        public static bool MoreThan<TSource>(
-            this IEnumerable<TSource> source,
-            int number)
-        {
-            Error.ThrowIfArgumentNull(nameof(source), source);
-            Error.ThrowIfArgumentSmallerThanZero(nameof(number), number);
-            return MoreThanInternal(source, number);
-        }
+    public static bool MoreThan<TSource>(
+        this IEnumerable<TSource> source,
+        int number)
+    {
+        Error.ThrowIfArgumentNull(nameof(source), source);
+        Error.ThrowIfArgumentSmallerThanZero(nameof(number), number);
+        return MoreThanInternal(source, number);
+    }
 
-        internal static bool MoreThanInternal<TSource>(
-            this IEnumerable<TSource> source,
-            int number)
+    internal static bool MoreThanInternal<TSource>(
+        this IEnumerable<TSource> source,
+        int number)
+    {
+        switch (source)
         {
-            if (source is ICollection<TSource> collectionT)
-            {
+            case ICollection<TSource> collectionT:
                 return collectionT.Count > number;
-            }
 #if !NET40
-            if (source is IReadOnlyCollection<TSource> readOnlyCollection)
-            {
+            case IReadOnlyCollection<TSource> readOnlyCollection:
                 return readOnlyCollection.Count > number;
-            }
 #endif
-            if (source is ICollection collection)
-            {
+            case ICollection collection:
                 return collection.Count > number;
-            }
-            int count = 0;
-            using var iterator = source.GetEnumerator();
-            while (iterator.MoveNext())
-            {
-                count += 1;
-                if (count > number)
-                {
-                    return true;
-                }
-            }
-            return false;
         }
+        int count = 0;
+        using var iterator = source.GetEnumerator();
+        while (iterator.MoveNext())
+        {
+            count += 1;
+            if (count > number)
+            {
+                return true;
+            }
+        }
+        return false;
     }
 }
