@@ -1,23 +1,21 @@
-﻿using System;
-using System.Text.RegularExpressions;
+﻿using System.Text.RegularExpressions;
 
-namespace RyanJuan.Hestia
+namespace RyanJuan.Hestia;
+
+public static partial class HestiaString
 {
-    public static partial class HestiaString
-    {
 #if ZH_HANT
 #else
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="value"></param>
+    /// <returns></returns>
 #endif
-        public static string RegexEscape(
-            this string value)
-        {
-            Error.ThrowIfArgumentNull(nameof(value), value);
-            return Regex.Escape(value);
-        }
+    public static string RegexEscape(
+        this string value)
+    {
+        Error.ThrowIfArgumentNull(nameof(value), value);
+        return Regex.Escape(value);
     }
 }
