@@ -1,25 +1,21 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+public static partial class HestiaCollections
 {
-    public static partial class HestiaCollections
-    {
 #if ZH_HANT
 #else
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="array"></param>
-        /// <param name="value"></param>
-        /// <returns></returns>
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="array"></param>
+    /// <param name="value"></param>
+    /// <returns></returns>
 #endif
-        public static bool Contains<TArray>(
-            this TArray[] array,
-            TArray value)
-        {
-            return Array.IndexOf(array, value) != -1;
-        }
+    [PublicAPI]
+    public static bool Contains<TArray>(
+        this TArray[] array,
+        TArray value)
+    {
+        return Array.IndexOf(array, value) != -1;
     }
 }
