@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+/// <summary>
+/// 
+/// </summary>
+public static partial class HestiaCollections
 {
-    /// <summary>
-    /// 
-    /// </summary>
-    public static partial class HestiaCollections
-    {
 #if ZH_HANT
 #else
 #endif
-    }
 }
