@@ -1,15 +1,29 @@
-﻿using System;
+namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+public static partial class HestiaEnum
 {
-    public static partial class HestiaEnum
-    {
 #if ZH_HANT
 #else
 #endif
-        public static TEnum[] GetValues<TEnum>()
-        {
-            return (TEnum[])EnumCacheCenter.GetValues(typeof(TEnum));
-        }
+    [PublicAPI]
+    public static
+#if NET40
+        IList<TEnum>
+#else
+        IReadOnlyList<TEnum>
+#endif
+        GetValues<TEnum>()
+        where TEnum : Enum
+    {
+        return EnumCacheCenter<TEnum>.Values;
+    }
+
+    [PublicAPI]
+    public static Array GetValues(Type type)
+    {
+        Error.ThrowIfArgumentNull(nameof(type), type);
+        ThrowIfTypeIsNotEnum(type);
+        var values = EnumCacheCenter.GetValues(type);
+        return (Array)values.Clone();
     }
 }
