@@ -1,18 +1,13 @@
-﻿using System;
-using System.Linq.Expressions;
-using System.Reflection;
+﻿namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+public static partial class HestiaReflection
 {
-    public static partial class HestiaReflection
-    {
 #if ZH_HANT
 #else
 #endif
-        public static object? GetDefaultValue(this Type type)
-        {
-            Error.ThrowIfArgumentNull(nameof(type), type);
-            return Activator.CreateInstance(type);
-        }
+    public static object? GetDefaultValue(this Type type)
+    {
+        Error.ThrowIfArgumentNull(nameof(type), type);
+        return Activator.CreateInstance(type);
     }
 }
