@@ -1,32 +1,28 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Reflection;
+namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+public static partial class ReflectionCenter
 {
-    public static partial class ReflectionCenter
+    public static T CreateInstance<T>()
     {
-        public static T CreateInstance<T>()
-        {
-            return ExpressionConstructor<T>.Default();
-        }
+        return ExpressionConstructor<T>.Factory.Invoke();
+    }
 
-        public static void SetInstanceFactory<T>(Func<T> factory)
-        {
-            Error.ThrowIfArgumentNull(nameof(factory), factory);
-            ExpressionConstructor<T>.Default = factory;
-        }
+    public static void SetInstanceFactory<T>(Func<T> factory)
+    {
+        Error.ThrowIfArgumentNull(nameof(factory), factory);
+        ExpressionConstructor<T>.Factory = factory;
+    }
 
-        private static class ExpressionConstructor<T>
-        {
-            private static Func<T>? s_default = null;
+    private static class ExpressionConstructor<T>
+    {
+        private static Func<T>? s_factory = null;
 
-            public static Func<T> Default
-            {
-                get => s_default ??= Expression.Lambda<Func<T>>(Expression.New(typeof(T))).Compile();
-                set => s_default = value;
-            }
+        public static Func<T> Factory
+        {
+            get => s_factory ??= Expression
+                .Lambda<Func<T>>(Expression.New(typeof(T)))
+                .Compile();
+            set => s_factory = value;
         }
     }
 }
