@@ -1,17 +1,10 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Reflection;
+﻿namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+public static partial class ReflectionCenter
 {
-    public static partial class ReflectionCenter
+    public static object? GetValue<T>(T instance, PropertyInfo property)
     {
-        public static object? GetValue<T>(T instance, PropertyInfo property)
-        {
-            Error.ThrowIfArgumentNull(nameof(property), property);
-            return ExpressionPropertyGetMethod<T>.GetGetMethod(property)(instance);
-        }
+        Error.ThrowIfArgumentNull(nameof(property), property);
+        return ExpressionPropertyGetMethod<T>.GetGetMethod(property).Invoke(instance);
     }
 }
