@@ -1,64 +1,60 @@
-﻿using System;
-using System.Diagnostics.CodeAnalysis;
+﻿namespace RyanJuan.Hestia;
 
-namespace RyanJuan.Hestia
+public static partial class HestiaString
 {
-    public static partial class HestiaString
-    {
 #if ZH_HANT
 #else
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="value"></param>
+    /// <returns></returns>
 #endif
-        public static bool IsWhiteSpace(
+    public static bool IsWhiteSpace(
 #if NETCOREAPP3_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-            [NotNullWhen(true)]
+        [NotNullWhen(true)]
 #endif
-            this string? value)
+        this string? value)
+    {
+        if (value == null)
         {
-            if (value == null)
+            return false;
+        }
+        for (int i = 0; i < value.Length; i += 1)
+        {
+            if (!char.IsWhiteSpace(value[i]))
             {
                 return false;
             }
-            for (int i = 0; i < value.Length; i++)
-            {
-                if (!char.IsWhiteSpace(value[i]))
-                {
-                    return false;
-                }
-            }
-            return true;
         }
+        return true;
+    }
 
 #if ZH_HANT
 #else
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="value"></param>
+    /// <returns></returns>
 #endif
-        public static bool IsNotWhiteSpace(
+    public static bool IsNotWhiteSpace(
 #if NETCOREAPP3_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-            [NotNullWhen(false)]
+        [NotNullWhen(false)]
 #endif
-            this string? value)
+        this string? value)
+    {
+        if (value == null)
         {
-            if (value == null)
-            {
-                return true;
-            }
-            for (int i = 0; i < value.Length; i++)
-            {
-                if (char.IsWhiteSpace(value[i]))
-                {
-                    return false;
-                }
-            }
             return true;
         }
+        for (int i = 0; i < value.Length; i += 1)
+        {
+            if (char.IsWhiteSpace(value[i]))
+            {
+                return false;
+            }
+        }
+        return true;
     }
 }
